rename post_factory_image to factory_image_c
