+++ backend/api/views/factory_image_c.py
